requirements.txt
# 云函数兼容版本 - 避免 pydantic_core 问题
fastapi==0.95.2
uvicorn==0.22.0
pydantic==1.10.12

# MySQL 数据库依赖
PyMySQL==1.1.0
aiomysql==0.2.0

# 其他依赖
python-multipart==0.0.6
python-socketio==5.11.0
texasholdem==0.11.0
